# Improve performance/allocations of WeakList